pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.5.3"
python-telegram-bot = "^13.7"
colorama = "^0.4.4"
SQLAlchemy = "^1.4.25"
alembic = "^1.7.3"
pandas = "^1.3.4"
psycopg2-binary = "^2.9.1"
python-dotenv = "^0.19.1"
requests = "^2.26.0"
Jinja2 = "^3.0.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
